+++ TablePro/Views/Results/Extensions/DataGridView+Click.swift
@@ -53,6 +53,7 @@ extension TableViewCoordinator {
             isTableEditable: isEditable,
             isRowDeleted: changeManager.isRowDeleted(row),
             isImmutableColumn: immutable.contains(columnName),
+            isBinaryValue: cellTypedValue(at: row, column: columnIndex).asBytes != nil,
             displayFormatOverride: override
         )
     }