+++ TableProTests/PluginTestSources/RedisArgumentCodec.swift
@@ -1,0 +1,1 @@
+../../Plugins/RedisDriverPlugin/RedisArgumentCodec.swift